+++ ruoyi-investment/src/main/java/com/ruoyi/investment/mapper/InvIndustryCsrcMapper.java
@@ -16,7 +16,7 @@ public interface InvIndustryCsrcMapper {
      * @param id 证监会行业主键
      * @return 证监会行业
      */
-    public InvIndustryCsrc selectInvIndustryCsrcById(Long id);
+    public InvIndustryCsrc selectInvIndustryCsrcById(Integer id);
 
     /**
      * 查询证监会行业列表
@@ -48,7 +48,7 @@ public interface InvIndustryCsrcMapper {
      * @param id 证监会行业主键
      * @return 结果
      */
-    public int deleteInvIndustryCsrcById(Long id);
+    public int deleteInvIndustryCsrcById(Integer id);
 
     /**
      * 批量删除证监会行业
