+++ ruoyi-investment/src/main/java/com/ruoyi/investment/service/IInvCompanyIndustryEmService.java
@@ -16,7 +16,7 @@ public interface IInvCompanyIndustryEmService {
      * @param id 所属东财行业主键
      * @return 所属东财行业
      */
-    public InvCompanyIndustryEm selectInvCompanyIndustryEmById(Long id);
+    public InvCompanyIndustryEm selectInvCompanyIndustryEmById(Integer id);
 
     /**
      * 查询所属东财行业列表
@@ -56,5 +56,5 @@ public interface IInvCompanyIndustryEmService {
      * @param id 所属东财行业主键
      * @return 结果
      */
-    public int deleteInvCompanyIndustryEmById(Long id);
+    public int deleteInvCompanyIndustryEmById(Integer id);
 }
