+++ ruoyi-investment/src/main/java/com/ruoyi/investment/service/IInvIndustryEmService.java
@@ -16,7 +16,7 @@ public interface IInvIndustryEmService {
      * @param id 东财行业主键
      * @return 东财行业
      */
-    public InvIndustryEm selectInvIndustryEmById(Long id);
+    public InvIndustryEm selectInvIndustryEmById(Integer id);
 
     /**
      * 查询东财行业列表
@@ -56,5 +56,5 @@ public interface IInvIndustryEmService {
      * @param id 东财行业主键
      * @return 结果
      */
-    public int deleteInvIndustryEmById(Long id);
+    public int deleteInvIndustryEmById(Integer id);
 }
